finished slider indicator and refactored carousel to reuse for product view
